requirements.txt
python-socketio
requests
psutil
